[FSET-1064] Add In Progress sub statuses to Test Data Generator.
