+++ app/services/testdata/StatusGeneratorFactory.scala
@@ -25,6 +25,9 @@ object StatusGeneratorFactory {
     status match {
       case "REGISTERED" => RegisteredStatusGenerator
       case ApplicationStatuses.Created => CreatedStatusGenerator
+      case "IN_PROGRESS_PERSONAL_DETAILS" => InProgressPersonalDetailsStatusGenerator
+      case "IN_PROGRESS_SCHEME_PREFERENCES" => InProgressSchemePreferencesStatusGenerator
+      case "IN_PROGRESS_ASSISTANCE_DETAILS" => InProgressAssistanceDetailsStatusGenerator
       case ApplicationStatuses.Submitted => SubmittedStatusGenerator
       case ApplicationStatuses.OnlineTestInvited => OnlineTestInvitedStatusGenerator
       case ApplicationStatuses.OnlineTestStarted => OnlineTestStartedStatusGenerator
@@ -50,3 +53,4 @@ object StatusGeneratorFactory {
   }
   // scalastyle:on cyclomatic.complexity
 }
+
