+++ test/mocks/application/PersonalDetailsInMemoryRepository.scala
@@ -18,7 +18,7 @@ package mocks.application
 
 import mocks.InMemoryStorage
 import model.Exceptions.PersonalDetailsNotFound
-import model.PersistedObjects.{ PersonalDetails, PersonalDetailsWithUserId }
+import model.PersistedObjects.{PersonalDetails, PersonalDetailsWithUserId}
 import org.joda.time.LocalDate
 import repositories.application.PersonalDetailsRepository
 
@@ -40,6 +40,9 @@ object PersonalDetailsInMemoryRepository extends PersonalDetailsRepository with 
   override def update(applicationId: String, userId: String, pd: PersonalDetails): Future[Unit] = {
     super.update(applicationId, userId, PersonalDetails(pd.firstName, pd.lastName, pd.preferredName, pd.dateOfBirth, pd.aLevel, pd.stemLevel))
   }
+
+  override def update(appId: String, userId: String, personalDetails: PersonalDetails,
+             requiredStatuses: Seq[String], newApplicationStatus: String) = Future.successful(())
 
   override def notFound(applicationId: String) = throw new PersonalDetailsNotFound(applicationId)
